Move our FloatingPointComparison and IntegerComparison to own header for reuse
